requirements.txt
python-telegram-bot==21.7
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
greenlet>=3.0.0
ruff>=0.8.0
